split() list of files into a new array instead of extracting file names in place
That more than doubles the memory we need, but for some reason
make the job to run a lot faster for large queries.
/usr/bin/time output on mu2eMissingJobs with a query that returned 100
files, before and after the change:

0.13user 0.01system 0:00.65elapsed 23%CPU (0avgtext+0avgdata 53808maxresident)k
32inputs+16outputs (0major+3457minor)pagefaults 0swaps

0.13user 0.02system 0:00.44elapsed 36%CPU (0avgtext+0avgdata 53856maxresident)k
32inputs+16outputs (0major+3456minor)pagefaults 0swaps

so difference for small queries.  However for a query that returned 10^6 entries
before the change, the job was about 40% done (file with stdout contained 400k lines)
after 17 hours of running.  After the change:

3.46user 0.78system 2:48.14elapsed 2%CPU (0avgtext+0avgdata 1536992maxresident)k
32inputs+95704outputs (0major+70432minor)pagefaults 0swaps

i.e. it completed in under 3 minutes, and 400MB of memory is still
very affordable.